% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{checkMiRNAVersion}
\alias{checkMiRNAVersion}
\title{check the miRNA Version in miRBase}
\usage{
checkMiRNAVersion(miRNANames)
}
\arguments{
\item{miRNANames}{A character vector representing the miRNA names.}
}
\value{
The version information is printed in the console.
}
\description{
This function checks miRBase Version for a list of miRNA names
}
\examples{
data(miRNATest)
miRNANames=miRNATest$miRNA_Name
checkMiRNAVersion(miRNANames)

}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
