% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{getAllMiRNAs}
\alias{getAllMiRNAs}
\title{Get all miRNAs in the specified miRBase version}
\usage{
getAllMiRNAs(version = "v21", type = "all", species = "all")
}
\arguments{
\item{version}{A character value representing the specified miRBase version for
retrieval. Users can apply the function \strong{getAllVersionInfo()} to get the available miRNA version names.
The optional values are in below:\cr
"v6","v7_1","v8","v8_1","v8_2","v9","v9_1",\cr
"v9_2","v10","v10_1","v11","v12","v13","v14",\cr
"v15","v16","v17","v18","v19","v20","v21"\cr}

\item{type}{A character value representing the miRNA type for retrieval.
\itemize{
\item \strong{"precursor"}  \cr
\item \strong{"mature"} \cr
\item \strong{"all"} : precursor and mature \cr
}}

\item{species}{A character value representing the abbreviation of species.  Users can apply
the \strong{getAllSpecies()} function to get the available  abbreviation of species. If species is set
to \strong{"all"}, the miRNAs of all species will return.}
}
\value{
A data frame with three columns. The three columns are defined as below:
\itemize{
 \item \strong{AccessionID} \cr
 \item \strong{Name} \cr
 \item \strong{Sequence} \cr
 }
}
\description{
This function gets all miRNAs in the specified miRBase version.
}
\examples{
miRNAs=getAllMiRNAs(version="v21", type="all", species="hsa")
}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
