% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{getAllSpecies}
\alias{getAllSpecies}
\title{Get all species of miRNAs embodied in miRBase repository}
\usage{
getAllSpecies()
}
\value{
A dataframe
A data frame with two columns. The two columns are defined as below:
\itemize{
 \item \strong{Species} \cr
 \item \strong{FullName} \cr
 }
}
\description{
This fuction return a reference for all species of miRNAs including the
abbreviation and full name.
}
\examples{
allSpecies=getAllSpecies()
}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
