% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{goToMiRBase}
\alias{goToMiRBase}
\title{Open the miRBase webpage of the specified miRNA}
\usage{
goToMiRBase(AccessionID)
}
\arguments{
\item{AccessionID}{A character representing a single miRNA Accession ID.}
}
\value{
No values
}
\description{
This function locates the miRBase webpage of the specified miRNA
}
\examples{
#### 1. A step-loop
AccessionID1="MI0000447"
goToMiRBase(AccessionID1)

#### 2. A mature miRNA
AccessionID2="MIMAT0026477"
goToMiRBase(AccessionID2)
}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
