% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{miRNA_NameToAccession}
\alias{miRNA_NameToAccession}
\title{The miRBase miRNA Name with specified version to Accession ID}
\usage{
miRNA_NameToAccession(miRNANames, Version = "v21")
}
\arguments{
\item{miRNANames}{A character vector representing the source miRNA names needed to be convert.}

\item{Version}{A character value representing the version name corresponding the miRNANames.
The optional values are in below:\cr
"v6","v7_1","v8","v8_1","v8_2","v9","v9_1",\cr
"v9_2","v10","v10_1","v11","v12","v13","v14",\cr
"v15","v16","v17","v18","v19","v20","v21"\cr}
}
\value{
A nx2 data matrix. The number of row equals to input miRNA name. The two columns are defined as below:


 \strong{Column 1} : The input miRNA names.\cr
 \strong{Column 2} : The convert Accession ID.\cr
}
\description{
This function converts a group of any species' miRNA name to the Accession ID defined in miRBase.
}
\examples{
data(miRNATest)
miRNANames=miRNATest[,2]
checkMiRNAVersion(miRNANames)
result1=miRNA_NameToAccession(miRNANames,Version="v18")
result2=miRNA_AccessionToName(result1[,2],targetVersion="v21")
result3=miRNAVersionConvert(miRNANames,targetVersion="v21",exact=TRUE)

}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
