% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNATest}
\alias{miRNATest}
\title{Dataset: miRNA Test Dataset}
\format{A data frame}
\description{
\itemize{
 \item Accession : miRNA Accessions in miRBase
 \item miRNA_Name : miRNA names
}
}
\note{
This is a mixed miRNA name collection of some species for the examples' test in the package.
}
\examples{
data(miRNATest)
}
\keyword{datasets}
