% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{checkMiRNAAlive}
\alias{checkMiRNAAlive}
\title{Check the miRNA status(Alive or Dead)}
\usage{
checkMiRNAAlive(AccessionIDs, verbose = TRUE)
}
\arguments{
\item{AccessionIDs}{A character vector representing the miRNA Accession IDs in miRBase.}

\item{verbose}{Logical value. If true, the dead miRNAs will be printed the console.}
}
\value{
A  character vector of Accession IDs for all alive miRNAs. The names of the return vector are the position indexes in the input AccessionIDs.
}
\description{
This function checks the miRNA status (Alive or Dead) in the latest miRBase verison.
}
\examples{
data(miRNATest)
## The input is miRNA Accession IDs
AccessionIDs=miRNATest$AccessionID
alive_AccessionID1=checkMiRNAAlive(AccessionIDs)

##The input is miRNA names
miRNANames=miRNATest$miRNA_Name
version=checkMiRNAVersion(miRNANames,verbose = TRUE)
result=miRNA_NameToAccession(miRNANames,version = version)
AccessionIDs=result$AccessionID
alive_AccessionID2=checkMiRNAAlive(AccessionIDs)
}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
