% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{checkMiRNAFamliy}
\alias{checkMiRNAFamliy}
\title{Check the miRNA family}
\usage{
checkMiRNAFamliy(AccessionIDs)
}
\arguments{
\item{AccessionIDs}{A character vector representing the miRNA Accession IDs in miRBase.}
}
\value{
A data frame with four columns. The number of rows equal to the input AccessionIDs. The four columns are defined as below:
\itemize{
 \item  \strong{AccessionID} : The input miRNA accessions.\cr
 \item \strong{miRNAName_v21} : The miRNA names (version 21) corresponding to the AccessionID.\cr
 \item \strong{FamliyAccession} : The accession ID of the family .\cr
 \item \strong{Family} : The family name.\cr
}
}
\description{
This function checks the miRNA family for a list of miRNA Names.
}
\examples{
data(miRNATest)
## The input is miRNAAccession IDs
AccessionIDs=miRNATest$AccessionID
Family_Info1=checkMiRNAFamliy(AccessionIDs)

##The input is miRNA names
miRNANames=miRNATest$miRNA_Name
version=checkMiRNAVersion(miRNANames,verbose = TRUE)
result=miRNA_NameToAccession(miRNANames,version=version)
AccessionIDs=result$AccessionID
Family_Info2=checkMiRNAFamliy(AccessionIDs)

}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
