% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{getMiRNAHistory}
\alias{getMiRNAHistory}
\title{Obtain all versions information for a single specified miRNA.}
\usage{
getMiRNAHistory(AccessionID)
}
\arguments{
\item{AccessionID}{A character representing the single AccessionID.}
}
\value{
A 21X7 data frame including all the history information (Precursor, Mature, Sequence) of the specified miRNA.
 Each row represents a miRBase version.
}
\description{
This function return all versions information of a single specified miRNA.
}
\examples{
#####1,The input is a miRNA Name
name="hsa-miR-26b-5p"
result1=miRNAVersionConvert(name,targetVersion="v21",exact=TRUE)
AccessionID=result1[,3]
result2=getMiRNAHistory(AccessionID)

#####2,The input is miRNA Accession Id
AccessionID="MIMAT0000765"
result3=getMiRNAHistory(AccessionID)

}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
