% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{goTo_miRBase}
\alias{goTo_miRBase}
\title{Open the miRBase webpages of the specified miRNAs}
\usage{
goTo_miRBase(AccessionIDs)
}
\arguments{
\item{AccessionIDs}{A character vector representing the miRNA Accession IDs in miRBase.
We restict the number of queried miRNAs each time. The maximum number of the input miRNAs is 15.}
}
\value{
No values
}
\description{
This function redirects the miRBase webpage of the specified miRNAs
}
\examples{
#### 1. A step-loop
AccessionID1="MI0000447"
goTo_miRBase(AccessionID1)

#### 2. A mature miRNA
AccessionID2="MIMAT0026477"
goTo_miRBase(AccessionID2)

#### 3. A list of miRNAs
AccessionID3=miRNATest$AccessionID[1:10]
goTo_miRBase(AccessionID3)

}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
