% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNATest}
\alias{miRNATest}
\title{Dataset: miRNA Test Dataset}
\format{A data frame}
\description{
\itemize{
 \item {Clumn 1} :miRNA Accession
 \item {Clumn 2} :miRNA Name
}
}
\note{
This test miRNA data are extracted from TarBase 6.0. Duo to the TarBase policy, We only list the miRNA names in Tarbase 6.0 for our package test.
}
\examples{
data(miRNATest)
}
\references{
Vergoulis, Thanasis, et al. "TarBase 6.0: capturing the exponential growth of miRNA targets with experimental support." Nucleic acids research 40.D1 (2012): D222-D229.
}
\keyword{datasets}
