% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{miRNAVersionConvert}
\alias{miRNAVersionConvert}
\title{miRBase name version convert}
\usage{
miRNAVersionConvert(miRNANames, targetVersion = "v21", exact = TRUE)
}
\arguments{
\item{miRNANames}{A character vector representing the source miRNA names needed to be convert.}

\item{targetVersion}{A character value representing the target miRBase version name corresponding the source miRNA name.
The optional values are in below:\cr
"v6","v7_1","v8","v8_1","v8_2","v9","v9_1",\cr
"v9_2","v10","v10_1","v11","v12","v13","v14",\cr
"v15","v16","v17","v18","v19","v20","v21"\cr}

\item{exact}{logical value. If true, the result will be the most exactly matched result.
If FALSE, the result will include all the possible matched miRNA name. If one miRNA can match multiple names. All the names will be
 concatenated with "&".}
}
\value{
A nx3 data matrix. The number of row equals to input miRNA name. The three columns are defined as below:


 \strong{Column 1} : The original miRNA name.\cr
 \strong{Column 2} : The converted miRBase name (in specified version) corresponding to the original miRNA name.\cr
 \strong{Column 3} : The corresponding miRBase accession ID.
}
\description{
This function converts a group of any species' miRNA names (including precursor and mature miRNA) to a specified miRBase version if the miRNAs have been defined in miRBase.
}
\note{
Please note: Due to some miRNAs' name changing many times in history. Even if choose the third parameter "exact"=TRUE, maybe it may still have some miRNAs that cannot
match the unique name in target version. In order to return the accurate result as possible, We also concatenate the multiple matched miRNA name with "&". This is rare case but it happens sometimes.
}
\examples{
data(miRNATest)
miRNANames=miRNATest[,2]
result1=miRNAVersionConvert(miRNANames,targetVersion="v13",exact=TRUE)
result2=miRNAVersionConvert(miRNANames,targetVersion="v21",exact=TRUE)
result3=miRNAVersionConvert(miRNANames,targetVersion="v21",exact=FALSE)

}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
